Latest conventions
- add toolchain resolver
- turn off junit parallel testing:
Because it really doesn't make much positive difference if Gradle is running in parallel mode with a decent number of test forks. On the downside, turning it on can cause some time-sensitive tests to fail, because they get starved.

Gradle parallel is much better, as it takes into account core count!