JPMS compatible schema output paths (#223)
JPMS requires resources that are accessible from outside the module to be in unique and open packages.

To achieve this, we introduced a `--output-strategy` to control if schemas should be generated under a `directoryTree` or a `flatDirectory`.

The plugin now defaults to no `outputDirectoryName`, and runs the generator with `--output-strategy=directoryTree`. This results in schema files being generated alongside their source types, i.e. in the same package.

If the user configures an `outputDirectoryName`, then the generator is run with `--output-strategy=flatDirectory`.